ubiwhere_challenge/requirements.txt
Django
psycopg2-binary
djangorestframework
djangorestframework-gis
markdown
django-filter
gunicorn
django-cors-headers
djangorestframework_simplejwt